+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `NodeAnimatedRotation` - A function that creates a `NodeSpin` action from a rotation property. Animating the rotation directly is not normally possible, it can only be done by controlling the angular speed. This function makes it possible by converting a rotation property to an angular speed property.
 - Added more utility functions:
   -  `FXRUtility.box` - Creates an outline of a cuboid shape.
+  -  `FXRUtility.rect` - Creates a rectangle.
   -  `FXRUtility.ellipse` - Creates an ellipse.
   -  `FXRUtility.ellipsoid` - Creates three ellipses that form the outline of an ellipsoid.
   -  `FXRUtility.cylinder` - Creates an outline of a cylinder.